+++ src/translators/sql_translator.py
@@ -803,7 +803,7 @@ class SQLTranslator(base_translator.BaseTranslator):
                 f"last_n should be >=1 and <= {default_limit}.")
         return min(last_n, limit)
 
-    def _get_where_clause(self, entity_ids, from_date, to_date, fiware_sp='/',
+    def _get_where_clause(self, entity_ids, from_date, to_date, idPattern, fiware_sp='/',
                           geo_query=None, prefix=''):
         clauses = []
         where_clause = ""
@@ -830,6 +830,10 @@ class SQLTranslator(base_translator.BaseTranslator):
         else:
             # Match prefix of fiware service path
             clauses.append(" " + prefix + FIWARE_SERVICEPATH + " = ''")
+
+        if idPattern:
+            clauses.append(  " " + prefix + ENTITY_ID_COL  + " ~* '" + idPattern +  "($|.*)'")
+
         # TODO implement prefix also for geo_clause
         geo_clause = self._get_geo_clause(geo_query)
         if geo_clause:
@@ -917,6 +921,7 @@ class SQLTranslator(base_translator.BaseTranslator):
               last_n=None,
               limit=10000,
               offset=0,
+              idPattern=None,
               fiware_service=None,
               fiware_servicepath='/',
               geo_query: SlfQuery = None):
@@ -1079,6 +1084,7 @@ class SQLTranslator(base_translator.BaseTranslator):
             where_clause = self._get_where_clause(entity_ids,
                                                   from_date,
                                                   to_date,
+                                                  idPattern,
                                                   fiware_servicepath,
                                                   geo_query)
 
@@ -1148,6 +1154,7 @@ class SQLTranslator(base_translator.BaseTranslator):
                   to_date=None,
                   limit=10000,
                   offset=0,
+                  idPattern=None,
                   fiware_service=None,
                   fiware_servicepath='/'):
         if limit == 0:
@@ -1156,6 +1163,7 @@ class SQLTranslator(base_translator.BaseTranslator):
         where_clause = self._get_where_clause(None,
                                               from_date,
                                               to_date,
+                                              idPattern,
                                               fiware_servicepath,
                                               None)
 
@@ -1215,6 +1223,7 @@ class SQLTranslator(base_translator.BaseTranslator):
                          to_date=None,
                          limit=10000,
                          offset=0,
+                         idPattern=None,
                          fiware_service=None,
                          fiware_servicepath='/'):
         if limit == 0:
@@ -1247,10 +1256,11 @@ class SQLTranslator(base_translator.BaseTranslator):
                 select_clause = self._get_select_clause(lower_attr_names, None,
                                                         None, prefix=prefix)
                 where_clause_no_prefix = self._get_where_clause(
-                    entity_ids, from_date, to_date, fiware_servicepath, None)
+                    entity_ids, from_date, to_date, idPattern, fiware_servicepath, None)
                 where_clause = self._get_where_clause(entity_ids,
                                                       from_date,
                                                       to_date,
+                                                      idPattern,
                                                       fiware_servicepath,
                                                       None, prefix=prefix)
                 stmt += "select {select} " \
@@ -1306,6 +1316,7 @@ class SQLTranslator(base_translator.BaseTranslator):
                          to_date=None,
                          limit=10000,
                          offset=0,
+                         idPattern=None,
                          fiware_service=None,
                          fiware_servicepath=None):
         if limit == 0:
@@ -1331,6 +1342,7 @@ class SQLTranslator(base_translator.BaseTranslator):
         where_clause = self._get_where_clause(entity_ids,
                                               from_date,
                                               to_date,
+                                              idPattern,
                                               fiware_servicepath)
 
         limit = min(10000, limit)
@@ -1529,11 +1541,12 @@ class SQLTranslator(base_translator.BaseTranslator):
                                     fiware_servicepath=fiware_servicepath)
 
     def delete_entities(self, etype, eid=None, from_date=None, to_date=None,
-                        fiware_service=None, fiware_servicepath='/'):
+                        idPattern=None, fiware_service=None, fiware_servicepath='/'):
         table_name = self._et2tn(etype, fiware_service)
         where_clause = self._get_where_clause(eid,
                                               from_date,
                                               to_date,
+                                              idPattern,
                                               fiware_servicepath)
         op = "delete from {} {}".format(table_name, where_clause)
         try: